refactor(settings): 拆出外观页签字体/背景/主题包面板与同步页签提供方卡片/表单
- 新增 settings/appearance/：FontManagerPanel、BackgroundPanel、ThemePackagePalette
- 新增 settings/sync/：SyncProviderCard、SyncProviderForms（webdav/r2/ftp/sftp 表单）、syncTabTypes
- AppearanceTab.tsx 685 -> 367 行；SyncTab.tsx 555 -> 280 行
